Updated thresholds to floats

+++ Core/Src/config.c
@@ -2,13 +2,13 @@
 #include "SERVO.h"
 #include "hand.h"
 
-int minThreshold = 0;
-int middleThreshold = 0;
-int maxThreshold = 0;
-int StartPosition = 0;
+float minThreshold = 0;
+float middleThreshold = 0;
+float maxThreshold = 0;
+float StartPosition = 0;
 
 //Calculates the 3 thresholds based on peak clenched voltage and average resting voltage
-void CalculateThresholds(int clenchedVoltage, int restingVoltage){
+void CalculateThresholds(float clenchedVoltage, float restingVoltage){
     minThreshold = restingVoltage * 4096 / 3.3 + 100;
     middleThreshold = (clenchedVoltage-restingVoltage) * 4096 / 3.3 * 0.45;
     maxThreshold = (clenchedVoltage-restingVoltage) * 4096 / 3.3 * 0.7;
